♻️ Refactor : MSA 도입에 따른 공통 모듈 정보 game 모듈로 도입 및 import 수정 및 코드 정리
