+++ module-common/src/main/java/com/example/modulecommon/annotation/LoginUser.java
@@ -1,12 +1,12 @@
-package com.example.modulecommon.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface LoginUser {
-
-}
+//package com.example.modulecommon.annotation;
+//
+//import java.lang.annotation.ElementType;
+//import java.lang.annotation.Retention;
+//import java.lang.annotation.RetentionPolicy;
+//import java.lang.annotation.Target;
+//
+//@Target(ElementType.PARAMETER)
+//@Retention(RetentionPolicy.RUNTIME)
+//public @interface LoginUser {
+//
+//}
