+++ module-common/src/main/java/com/example/modulecommon/entity/AuthUser.java
@@ -1,11 +1,11 @@
-package com.example.modulecommon.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class AuthUser {
-	private final Long memberId;
-	private final String role;
-}
+//package com.example.modulecommon.entity;
+//
+//import lombok.Getter;
+//import lombok.RequiredArgsConstructor;
+//
+//@Getter
+//@RequiredArgsConstructor
+//public class AuthUser {
+//	private final Long memberId;
+//	private final String role;
+//}
