+++ module-common/src/main/java/com/example/modulecommon/entity/Timestamped.java
@@ -1,28 +1,28 @@
-package com.example.modulecommon.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class Timestamped {
-	
-	@CreatedDate
-	@Column(updatable = false)
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime createdAt;
-	
-	@LastModifiedDate
-	@Column
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime updatedAt;
-}
+//package com.example.modulecommon.entity;
+//
+//import jakarta.persistence.Column;
+//import jakarta.persistence.EntityListeners;
+//import jakarta.persistence.MappedSuperclass;
+//import jakarta.persistence.Temporal;
+//import jakarta.persistence.TemporalType;
+//import java.time.LocalDateTime;
+//import lombok.Getter;
+//import org.springframework.data.annotation.CreatedDate;
+//import org.springframework.data.annotation.LastModifiedDate;
+//import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+//
+//@Getter
+//@MappedSuperclass
+//@EntityListeners(AuditingEntityListener.class)
+//public abstract class Timestamped {
+//
+//	@CreatedDate
+//	@Column(updatable = false)
+//	@Temporal(TemporalType.TIMESTAMP)
+//	private LocalDateTime createdAt;
+//
+//	@LastModifiedDate
+//	@Column
+//	@Temporal(TemporalType.TIMESTAMP)
+//	private LocalDateTime updatedAt;
+//}
