+++ module-common/src/main/java/com/example/modulecommon/exception/ServerException.java
@@ -1,17 +1,17 @@
-package com.example.modulecommon.exception;
-
-import lombok.Getter;
-
-@Getter
-public class ServerException extends RuntimeException {
-	private final ErrorCode errorCode;
-
-	public ServerException(ErrorCode errorCode) {
-		super(errorCode.getMessage());
-		this.errorCode = errorCode;
-	}
-
-	public ErrorCode getErrorCode() {
-		return errorCode;
-	}
-}
+//package com.example.modulecommon.exception;
+//
+//import lombok.Getter;
+//
+//@Getter
+//public class ServerException extends RuntimeException {
+//	private final ErrorCode errorCode;
+//
+//	public ServerException(ErrorCode errorCode) {
+//		super(errorCode.getMessage());
+//		this.errorCode = errorCode;
+//	}
+//
+//	public ErrorCode getErrorCode() {
+//		return errorCode;
+//	}
+//}
