+++ module-common/src/main/java/com/example/modulecommon/util/PageUtil.java
@@ -1,10 +1,10 @@
-package com.example.modulecommon.util;
-
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-
-public class PageUtil {
-	public static Pageable correctPageIndex(Pageable pageable){
-		return PageRequest.of(pageable.getPageNumber() - 1, pageable.getPageSize());
-	}
-}
+//package com.example.modulecommon.util;
+//
+//import org.springframework.data.domain.PageRequest;
+//import org.springframework.data.domain.Pageable;
+//
+//public class PageUtil {
+//	public static Pageable correctPageIndex(Pageable pageable){
+//		return PageRequest.of(pageable.getPageNumber() - 1, pageable.getPageSize());
+//	}
+//}
