+++ module-game/src/main/java/com/example/modulegame/common/image/ImageService.java
@@ -3,8 +3,8 @@ package com.example.modulegame.common.image;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.example.modulecommon.exception.ErrorCode;
-import com.example.modulecommon.exception.ServerException;
+import com.example.modulegame.global.exception.ErrorCode;
+import com.example.modulegame.global.exception.ServerException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
