--- module-game/src/main/java/com/example/modulegame/config/RSocketClientConfig.java
@@ -1,15 +1,0 @@
-package com.example.modulegame.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.rsocket.RSocketRequester;
-
-//@Configuration
-//public class RSocketClientConfig {
-//
-//    @Bean
-//    public RSocketRequester rSocketRequester(RSocketRequester.Builder builder) {
-//        return builder
-//                .tcp("localhost", 7000); // 티켓 서버 주소
-//    }
-//}
