+++ module-game/src/main/java/com/example/modulegame/domain/game/controller/GameController.java
@@ -102,5 +102,4 @@ public class GameController {
     ) {
         return ResponseEntity.ok(gameService.getGameDtoList(gameIds));
     }
-
 }
