+++ module-game/src/main/java/com/example/modulegame/domain/stadium/repository/SectionRepository.java
@@ -1,12 +1,10 @@
 package com.example.modulegame.domain.stadium.repository;
 
-
 import com.example.modulegame.domain.stadium.entity.Section;
 import com.example.modulegame.domain.stadium.entity.Stadium;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SectionRepository extends JpaRepository<Section, Long> {
-
 
     boolean existsByCodeAndStadium(String code, Stadium stadium);
 }
