+++ module-game/src/main/java/com/example/modulegame/domain/stadium/repository/StadiumRepository.java
@@ -1,11 +1,9 @@
 package com.example.modulegame.domain.stadium.repository;
 
-
 import com.example.modulegame.domain.stadium.entity.Stadium;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StadiumRepository extends JpaRepository<Stadium, Long> {
-
 
     boolean existsByName(String name);
