+++ module-game/src/main/java/com/example/modulegame/domain/stadium/service/SectionService.java
@@ -1,7 +1,5 @@
 package com.example.modulegame.domain.stadium.service;
 
-
-import com.example.modulecommon.exception.ErrorCode;
 import com.example.modulegame.domain.stadium.dto.request.SectionCreateRequest;
 import com.example.modulegame.domain.stadium.dto.request.SectionUpdateRequest;
 import com.example.modulegame.domain.stadium.dto.response.SectionCreateResponse;
@@ -9,10 +7,11 @@ import com.example.modulegame.domain.stadium.dto.response.SectionUpdateResponse;
 import com.example.modulegame.domain.stadium.entity.Section;
 import com.example.modulegame.domain.stadium.entity.Stadium;
 import com.example.modulegame.domain.stadium.repository.SectionRepository;
+import com.example.modulegame.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import com.example.modulecommon.exception.ServerException;
+import com.example.modulegame.global.exception.ServerException;
 
 @Service
 @RequiredArgsConstructor
