+++ module-game/src/main/java/com/example/modulegame/feign/client/TicketClient.java
@@ -4,8 +4,6 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-import java.util.Set;
-
 @FeignClient(name = "ticket", url = "http://localhost:8082")
 public interface TicketClient {
