+++ module-game/src/main/java/com/example/modulegame/global/annotation/LoginUser.java
@@ -1,0 +1,12 @@
+package com.example.modulegame.global.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginUser {
+
+}
