+++ module-game/src/main/java/com/example/modulegame/global/argumentresolver/AuthUserArgumentResolver.java
@@ -1,0 +1,28 @@
+package com.example.modulegame.global.argumentresolver;
+
+import com.example.modulegame.global.annotation.LoginUser;
+import com.example.modulegame.global.entity.AuthUser;
+import org.springframework.core.MethodParameter;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
+
+@Component
+public class AuthUserArgumentResolver implements HandlerMethodArgumentResolver {
+
+	@Override
+	public boolean supportsParameter(MethodParameter parameter) {
+		return parameter.hasParameterAnnotation(LoginUser.class) && parameter.getParameterType().equals(AuthUser.class);
+	}
+
+	@Override
+	public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
+		NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
+
+		Long memberId = Long.valueOf(webRequest.getHeader("member-id"));
+		String memberRole = webRequest.getHeader("member-role");
+		return new AuthUser(memberId, memberRole);
+	}
+}
