+++ module-game/src/main/java/com/example/modulegame/global/entity/AuthUser.java
@@ -1,0 +1,11 @@
+package com.example.modulegame.global.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class AuthUser {
+	private final Long memberId;
+	private final String role;
+}
