+++ module-game/src/main/java/com/example/modulegame/global/entity/Timestamped.java
@@ -1,0 +1,25 @@
+package com.example.modulegame.global.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class Timestamped {
+	
+	@CreatedDate
+	@Column(updatable = false)
+	@Temporal(TemporalType.TIMESTAMP)
+	private LocalDateTime createdAt;
+	
+	@LastModifiedDate
+	@Column
+	@Temporal(TemporalType.TIMESTAMP)
+	private LocalDateTime updatedAt;
+}
