+++ module-game/src/main/java/com/example/modulegame/global/exception/ErrorResponse.java
@@ -1,9 +1,10 @@
-package com.example.modulegame.common.exception;
+package com.example.modulegame.global.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 
 import java.util.HashMap;
 import java.util.Map;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
