+++ module-game/src/main/java/com/example/modulegame/global/exception/GlobalExceptionHandler.java
@@ -1,16 +1,15 @@
-package com.example.modulegame.common.exception;
+package com.example.modulegame.global.exception;
 
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
-
-import com.example.modulecommon.exception.ErrorCode;
-import com.example.modulecommon.exception.ServerException;
-import java.util.HashMap;
-import java.util.Map;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
