+++ module-game/src/main/java/com/example/modulegame/global/util/PageUtil.java
@@ -1,0 +1,10 @@
+package com.example.modulegame.global.util;
+
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+
+public class PageUtil {
+	public static Pageable correctPageIndex(Pageable pageable){
+		return PageRequest.of(pageable.getPageNumber() - 1, pageable.getPageSize());
+	}
+}
